+++ tests/testthat/test-new-dataset.R
@@ -11,6 +11,17 @@ test_that("newDataset input validation", {
         "Can only make a Crunch dataset from a two-dimensional data")
 })
 
+with_mock_HTTP({
+    test_that("Basic exercise of turning data.frame to Crunch payload", {
+        expect_POST(newDataset(data.frame(a=1), name="Testing"),
+            "/api/datasets/",
+            '{"element":"shoji:entity","body":{"name":"Testing",',
+            '"table":{"element":"crunch:table",',
+            '"metadata":{"a":{"type":"numeric","name":"a","alias":"a"}},',
+            '"order":["a"]}}}')
+    })
+})
+
 if (run.integration.tests) {
     test_that("Source file cannot be uploaded if not logged in", {
         logout()
@@ -47,10 +58,15 @@ if (run.integration.tests) {
 
         test_that("createWithMetadataAndFile using docs example", {
             ds <- newDatasetFromFixture("apidocs")
-            expect_true(is.dataset(ds))
-            expect_identical(name(ds), "Example dataset")
-            expect_identical(names(categories(ds$q1)),
-                c("Cat", "Dog", "Bird", "Skipped", "Not Asked"))
+            expect_valid_apidocs_import(ds)
+        })
+
+        test_that("data.frame with spaces in column names", {
+            input <- data.frame(a=factor("A"), b=4)
+            names(input) <- c("var one", "var two")
+            expect_identical(names(input), c("var one", "var two"))
+            ds <- newDataset(input)
+            expect_identical(names(ds), c("var one", "var two"))
         })
 
         purgeEntitiesCreated()