+++ NEWS.md
@@ -1,7 +1,10 @@
 ### crunch 1.11.99 (under development)
 * **Breaking change**: Accessing subvariables from array variables is now done by alias, just as variables are extracted from a Dataset. The "crunch.namekey.dataset" and "crunch.namekey.array" options have existed for a while, but they've had different default values. Now both default to "alias", which should offer a more consistent interface. If you want to maintain the old behavior, you can set `options(crunch.namekey.array="name")` in your script or in your .Rprofile.
 * `deleteSubvariable` now follows "crunch.namekey.array" and will take either subvariable names or aliases, depending on the value of the setting.
+* New `extendDataset` function, also aliased as a `merge` method, to allow you to add columns from one dataset to another, joining on a key variable from each.
 * Attempting to assign a `name<-` on `NULL` (i.e. when you reference a variable in a dataset using `$` and the variable does not exist) returns a helpful message.
+* Fix dataset import via `newDataset` when passing a `data.frame` or similar that has spaces in the column names.
+* `notes` and `notes<-` methods for datasets, variables, and variable catalogs to view and edit those new metadata fields.
 
 ### crunch 1.11.2
 * Fix issue with sharing datasets owned by a project.